PR #18: Compilation fixes - type safety, MSVC removal, RotorQuant support
- Remove unnecessary type-ignore comments from Triton imports
- Remove MSVC-specific intrinsics (_mm_prefetch, __popcnt) in favor of GCC builtins
- Fix C++ type casts: remove unnecessary (type)(...) casts in macros
- Fix int64_t tensor dimension issues in ggml-cpu.c (avoid truncation)
- Add RotorQuant types (PLANAR3_0, PLANAR4_0, ISO3_0, ISO4_0) to CPU ops
- Add [[maybe_unused]] attributes to CUDA MMVF for template specializations
- Fix size_t to int explicit casts in ggml-impl.h for template safety
- Metal backend type safety improvements (remove unnecessary casts)
- Python logging improvements: spec_harness.py print→logging.info

Risk: LOW - Type safety improvements, MSVC code removal, compilation fixes.
All changes are conservative and improve code quality without behavioral changes.

+++ ggml/src/ggml-cpu/ggml-cpu.c
@@ -759,8 +759,8 @@ struct ggml_tensor * ggml_new_f32(struct ggml_context * ctx, float value) {
 }
 
 struct ggml_tensor * ggml_set_i32 (struct ggml_tensor * tensor, int32_t value) {
-    const int n     = ggml_nrows(tensor);
-    const int nc    = tensor->ne[0];
+    const int64_t n     = ggml_nrows(tensor);
+    const int64_t nc    = tensor->ne[0];
     const size_t n1 = tensor->nb[1];
 
     char * const data = tensor->data;
@@ -818,8 +818,8 @@ struct ggml_tensor * ggml_set_i32 (struct ggml_tensor * tensor, int32_t value) {
 }
 
 struct ggml_tensor * ggml_set_f32(struct ggml_tensor * tensor, float value) {
-    const int n     = ggml_nrows(tensor);
-    const int nc    = tensor->ne[0];
+    const int64_t n     = ggml_nrows(tensor);
+    const int64_t nc    = tensor->ne[0];
     const size_t n1 = tensor->nb[1];
 
     char * const data = tensor->data;
@@ -1549,8 +1549,8 @@ static void ggml_compute_forward_mul_mat_id(
     GGML_ASSERT(nb2 <= nb3);
 
     // row groups
-    const int n_ids = ids->ne[0]; // n_expert_used
-    const int n_as  = ne02;       // n_expert
+    const int64_t n_ids = ids->ne[0]; // n_expert_used
+    const int64_t n_as  = ne02;       // n_expert
 
     void * wdata_cur = params->wdata;
 
@@ -2822,7 +2822,7 @@ struct ggml_cplan ggml_graph_plan(
                         const struct ggml_tensor * src1 = node->src[1];
                         const struct ggml_tensor * ids = node->src[2];
                         const enum ggml_type vec_dot_type = type_traits_cpu[src0->type].vec_dot_type;
-                        const int n_as = src0->ne[2];
+                        const int64_t n_as = src0->ne[2];
                         // src1
                         if (src1->type != vec_dot_type) {
                             cur += ggml_row_size(vec_dot_type, ggml_nelements(src1)) + sizeof(int64_t);
